requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==21.2.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
